+++ src/library/adash-messages.adb
@@ -581,6 +581,8 @@ package body Adash.Messages is
             return "command.pipe_start.doc";
          when Msg_Command_Pipe_From_Doc =>
             return "command.pipe_from.doc";
+         when Msg_Command_Pipe_From_Text_Doc =>
+            return "command.pipe_from_text.doc";
          when Msg_Command_Pipe_Into_Doc =>
             return "command.pipe_into.doc";
          when Msg_Command_Pipe_Append_Doc =>
@@ -985,6 +987,7 @@ package body Adash.Messages is
             | Msg_Command_Run_All_New_Doc
             | Msg_Command_Pipe_Doc
             | Msg_Command_Pipe_Start_Doc
+            | Msg_Command_Pipe_From_Text_Doc
             | Msg_Command_Pipe_From_Doc
             | Msg_Command_Pipe_Into_Doc
             | Msg_Command_Pipe_Append_Doc